v0.8.7

+++ src/update.command
@@ -49,12 +49,21 @@ CHECK_DOCKER_RC=$(echo $DOCKER_VERSION | grep rc)
 if [ -n "$CHECK_DOCKER_RC" ]
 then
     # docker RC release
-    echo "Downloading docker $DOCKER_VERSION client for OS X"
-    curl -o ~/coreos-osx/bin/docker http://test.docker.com/builds/Darwin/x86_64/docker-$DOCKER_VERSION
+    if [ -n "curl -s --head https://test.docker.com/builds/Darwin/x86_64/docker-$DOCKER_VERSION | head -n 1 | grep "HTTP/1.[01] [23].." | grep 200" ]
+    then
+        # we check if RC is still available
+        echo "Downloading docker $DOCKER_VERSION client for OS X"
+        curl -o ~/coreos-osx/bin/docker http://test.docker.com/builds/Darwin/x86_64/docker-$DOCKER_VERSION
+    else
+        # RC is not available anymore, so we download stable release
+        DOCKER_VERSION_STABLE=$(echo $DOCKER_VERSION | cut -d"-" -f1)
+        echo "Downloading docker $DOCKER_VERSION_STABLE client for OS X"
+        curl -o ~/coreos-osx/bin/docker http://get.docker.com/builds/Darwin/x86_64/docker-$DOCKER_VERSION_STABLE
+    fi
 else
     # docker stable release
     echo "Downloading docker $DOCKER_VERSION client for OS X"
-    curl -o ~/coreos-osx/bin/docker http://get.docker.io/builds/Darwin/x86_64/docker-$DOCKER_VERSION
+    curl -o ~/coreos-osx/bin/docker http://get.docker.com/builds/Darwin/x86_64/docker-$DOCKER_VERSION
 fi
 # Make it executable
 chmod +x ~/coreos-osx/bin/docker
